feat: support narrower integer fields (i32/i16/i8) over INT64
sea-orm's proxy read is strict: i32::try_from(Value::BigInt) fails, so a
model field declared as i32 over a Spanner INT64 column could not be read
(it surfaced as 'Missing value for column'). No single Value variant can
satisfy both i32 and i64, and the orphan rule blocks fixing it in-crate.

Fix it upstream via lenient integer coercion in sea-query, consumed here by:
- bumping sea-orm rc.37 -> rc.41 (whose sea-query req relaxed to ~1.0.0)
  and sea-query rc.31 -> 1.0; MSRV is now 1.94 (already effectively 1.85).
- a temporary [patch.crates-io] onto the sea-query PR branch
  (devgony/sea-query lenient-integer-value-type) until it is released.
- implementing the new required QueryBuilder::prepare_explain_statement
  (no-op; Spanner doesn't EXPLAIN through this builder).

all_types now declares int32_val/int32_nullable as i32, exercised by the
existing type tests. Also relocates proxy unit tests to end-of-file to
satisfy clippy::items_after_test_module on the newer toolchain.

Remove the patch once the coercion ships in a released sea-query 1.0.x.

+++ tests/entity/all_types.rs
@@ -11,8 +11,9 @@ pub struct Model {
 
     pub int64_val: i64,
     pub int64_nullable: Option<i64>,
-    pub int32_val: i64,
-    pub int32_nullable: Option<i64>,
+    // i32 over an INT64 column: exercises lenient integer read coercion.
+    pub int32_val: i32,
+    pub int32_nullable: Option<i32>,
 
     pub float64_val: f64,
     pub float64_nullable: Option<f64>,
